chore: extract the common used component to package init file
